moved couple things around
