+++ src/lvl_script_commands.c
@@ -25,11 +25,12 @@
 #include "game_merge.h"
 #include "game_legacy.h"
 #include "music_player.h"
-#include "lvl_script_lib.h"
-#include "lvl_script_conditions.h"
 #include "keeperfx.hpp"
 
 #include "lvl_script_value.h"
+#include "lvl_script_lib.h"
+#include "lvl_script_conditions.h"
+#include "lvl_script_parser.h"
 
 #ifdef __cplusplus
 extern "C" {
