+++ src/lvl_script_commands.h
@@ -23,6 +23,7 @@ extern "C" {
 #endif
 
 #include "lvl_script_lib.h"
+#include "lvl_script_parser.h"
 
 void script_add_command(const struct CommandDesc *cmd_desc, const struct ScriptLine *scline);
