+++ src/lvl_script_lib.h
@@ -249,26 +249,7 @@ enum ScriptVariables {
   SVar_BONUS_TIME                      = 76,
  };
 
-extern const struct CommandDesc command_desc[];
 
-extern const struct CommandDesc subfunction_desc[];
-extern const struct NamedCommand newcrtr_desc[];
-extern const struct NamedCommand player_desc[];
-//extern const struct NamedCommand variable_desc[];
-extern const struct NamedCommand controls_variable_desc[];
-extern const struct NamedCommand comparison_desc[];
-extern const struct NamedCommand head_for_desc[];
-extern const struct NamedCommand timer_desc[];
-extern const struct NamedCommand flag_desc[];
-extern const struct NamedCommand hero_objective_desc[];
-extern const struct NamedCommand msgtype_desc[];
-extern const struct NamedCommand tendency_desc[];
-extern const struct NamedCommand creature_select_criteria_desc[];
-extern const struct NamedCommand door_config_desc[];
-extern const struct NamedCommand trap_config_desc[];
-extern const struct NamedCommand gui_button_group_desc[];
-extern const struct NamedCommand campaign_flag_desc[];
-extern const struct NamedCommand script_operator_desc[];
 
 
 
@@ -278,14 +259,10 @@ extern const struct NamedCommand script_operator_desc[];
 
 
 long parse_creature_name(const char *creature_name);
-#define get_map_location_id(locname, location) get_map_location_id_f(locname, location, __func__, text_line_number)
-TbBool get_map_location_id_f(const char *locname, TbMapLocation *location, const char *func_name, long ln_num);
 struct ScriptValue *allocate_script_value(void);
 struct Thing *script_process_new_object(long crmodel, TbMapLocation location, long arg);
 void command_init_value(struct ScriptValue* value, unsigned long var_index, unsigned long plr_range_id);
 void command_add_value(unsigned long var_index, unsigned long plr_range_id, long val2, long val3, long val4);
-#define get_map_heading_id(headname, target, location) get_map_heading_id_f(headname, target, location, __func__, text_line_number)
-TbBool get_map_heading_id_f(const char *headname, long target, TbMapLocation *location, const char *func_name, long ln_num);
 #define get_players_range(plr_range_id, plr_start, plr_end) get_players_range_f(plr_range_id, plr_start, plr_end, __func__, text_line_number)
 long get_players_range_f(long plr_range_id, int *plr_start, int *plr_end, const char *func_name, long ln_num);
 TbBool parse_set_varib(const char *varib_name, long *varib_id, long *varib_type);
