+++ src/lvl_script_parser.h
@@ -22,8 +22,30 @@
 extern "C" {
 #endif
 
-int get_script_current_condition();
-void set_script_current_condition(int current_condition);
+
+#include "lvl_script_lib.h"
+
+extern const struct CommandDesc command_desc[];
+extern const struct CommandDesc dk1_command_desc[];
+
+
+extern const struct CommandDesc subfunction_desc[];
+extern const struct NamedCommand newcrtr_desc[];
+extern const struct NamedCommand player_desc[];
+//extern const struct NamedCommand variable_desc[];
+extern const struct NamedCommand controls_variable_desc[];
+extern const struct NamedCommand comparison_desc[];
+extern const struct NamedCommand timer_desc[];
+extern const struct NamedCommand flag_desc[];
+extern const struct NamedCommand hero_objective_desc[];
+extern const struct NamedCommand msgtype_desc[];
+extern const struct NamedCommand tendency_desc[];
+extern const struct NamedCommand creature_select_criteria_desc[];
+extern const struct NamedCommand door_config_desc[];
+extern const struct NamedCommand trap_config_desc[];
+extern const struct NamedCommand gui_button_group_desc[];
+extern const struct NamedCommand campaign_flag_desc[];
+extern const struct NamedCommand script_operator_desc[];
 
 #ifdef __cplusplus
 }
